+++ rmk/CHANGELOG.md
@@ -11,6 +11,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix compilation error when use `Macro()` in keymap config
 
+### Added
+
+- Add PMW3610 optical mouse sensor support for nRF and RP2040 with bit-bang SPI
+
+### Changed
+
+- Make `embedded-hal-async` a required dependency
+- Update default BLE connection parameters
+
 ## [0.8.1] - 2025-11-25
 
 ### Changed